Modify noncustom xsl to work with oral history
CDM_Util conversion uses cdm_to_foxml_noncustom in lieu of a CDM to Oral
History XSL. It apperaws that the non-custom XSL has not been called on
any of the existing CONTENTdm collections. Had to Modify output file
generator so it is not dependent on the presence of a base-uri. This
is the same method used by implemented and tested xsl files.